Отдельный контроллер для сотрудников.
