+++ WebStore/Views/Employees/EmployeeDetails.cshtml
@@ -9,6 +9,6 @@
         <p>Возраст: @Model.Age</p>
 
         <hr />
-        <a href="/Home/Index">К списку сотирудников</a>
+        <a asp-controller="Employees" asp-action="Index">К списку сотирудников</a>
 
     </div>
