+++ WebStore/Views/Employees/Index.cshtml
@@ -1,0 +1,32 @@
+﻿@model IEnumerable<Employee>
+@{
+    ViewData["Title"] = "Index";
+}
+
+<div class="container">
+    <table class="table">
+        <thead>
+            <tr>
+                <th>id</th>
+                <th>Фамилия</th>
+                <th>Имя</th>
+                <th>Отчество</th>
+                <th>Возраст</th>
+                <th></th>
+            </tr>
+        </thead>
+        <tbody>
+            @foreach(var empl in Model)
+            {
+                <tr>
+                    <td>@empl.Id</td>
+                    <td>@empl.Surname</td>
+                    <td>@empl.FirstName</td>
+                    <td>@empl.Patronymic</td>
+                    <td>@empl.Age</td>
+                    <td><a asp-controller="Employees" asp-action="EmployeeDetails" asp-route-id="@empl.Id" >Информация</a></td>
+                </tr>
+            }
+        </tbody>
+    </table>
+</div>
